chore(docutils): break out fs-related functions in util into their own module
